requirements.txt
fastapi
uvicorn[standard]
python-multipart
supabase
python-dotenv
pydantic
